llama3/requirements.txt
streamlit
groq
python-dotenv